requirements.txt
numpy==1.21.2
streamlit==0.88.0
opencv_python==4.5.3.56
mediapipe==0.8.8
Pillow==8.3.2
